+++ StereoKitC/asset_types/font.cpp
@@ -53,13 +53,13 @@ void         font_source_release (int32_t id);
 
 int32_t font_source_add(const char *filename) {
 	int64_t hash = hash_fnv64_string(filename);
-	int32_t id   = (int32_t)font_sources.index_where(&font_source_t::name_hash, hash);
+	int32_t id   = font_sources.index_where(&font_source_t::name_hash, hash);
 
 	if (id == -1) {
 		font_source_t new_file = {};
 		new_file.name_hash = hash;
 		new_file.name      = string_copy(filename);
-		id = (int32_t)font_sources.add(new_file);
+		id = font_sources.add(new_file);
 	}
 	font_sources[id].references += 1;
 
@@ -85,13 +85,13 @@ int32_t font_source_add(const char *filename) {
 
 int32_t font_source_add_data(const char *name, const void *data, size_t data_size) {
 	int64_t hash = hash_fnv64_string(name);
-	int32_t id   = (int32_t)font_sources.index_where(&font_source_t::name_hash, hash);
+	int32_t id   = font_sources.index_where(&font_source_t::name_hash, hash);
 
 	if (id == -1) {
 		font_source_t new_file = {};
 		new_file.name_hash = hash;
 		new_file.name      = string_copy(name);
-		id = (int32_t)font_sources.add(new_file);
+		id = font_sources.add(new_file);
 	}
 	font_sources[id].references += 1;
 
@@ -260,11 +260,11 @@ void font_release(font_t font) {
 ///////////////////////////////////////////
 
 void font_destroy(font_t font) {
-	int64_t idx = font_list.index_of(font);
+	int32_t idx = font_list.index_of(font);
 	if (idx >= 0)
-		font_list.remove((size_t)idx);
+		font_list.remove(idx);
 
-	for (size_t i = 0; i < font->font_ids.count; i++) {
+	for (int32_t i = 0; i < font->font_ids.count; i++) {
 		font_source_release(font->font_ids[i]);
 	}
 
@@ -402,7 +402,7 @@ void font_upsize_texture(font_t font) {
 	// size. We'll copy glyph by glyph to make sorting easier later on.
 	uint8_t *new_data = sk_malloc_t(uint8_t, new_w*new_h);
 	memset(new_data, 0, new_w * new_h);
-	for (size_t i = 0; i < font->glyph_map.items.count; i++) {
+	for (int32_t i = 0; i < font->glyph_map.items.count; i++) {
 		font_char_t ch     = font->glyph_map.items[i];
 		font_char_t new_ch = ch;
 		font_char_reuv(&new_ch, scale_x, scale_y);
@@ -419,12 +419,12 @@ void font_upsize_texture(font_t font) {
 		}
 	}
 	// Update the rest of our rectangles
-	for (size_t i = 32; i < 128;                             i++) font_char_reuv(&font->characters         [i], scale_x, scale_y);
-	for (size_t i = 0;  i < font->character_map.items.count; i++) font_char_reuv(&font->character_map.items[i], scale_x, scale_y);
+	for (int32_t i = 32; i < 128;                             i++) font_char_reuv(&font->characters         [i], scale_x, scale_y);
+	for (int32_t i = 0;  i < font->character_map.items.count; i++) font_char_reuv(&font->character_map.items[i], scale_x, scale_y);
 	
 	// This could be a faster copy, but may not make a big difference. Also
 	// doesn't allow for copying to new locations.
-	/*for (size_t i = 0;  i < font->glyph_map.items.count; i++) font_char_reuv(&font->glyph_map.items[i], scale_x, scale_y);
+	/*for (int32_t i = 0;  i < font->glyph_map.items.count; i++) font_char_reuv(&font->glyph_map.items[i], scale_x, scale_y);
 	for (int32_t y = 0; y < font->atlas.h; y++) {
 		memcpy(&new_data[y * new_w], &font->atlas_data[y * font->atlas.w], font->atlas.w * sizeof(uint8_t));
 	}*/
@@ -448,7 +448,7 @@ void font_update_texture(font_t font) {
 ///////////////////////////////////////////
 
 font_glyph_t font_find_glyph(font_t font, char32_t character) {
-	for (size_t i = 0; i < font->font_ids.count; i++) {
+	for (int32_t i = 0; i < font->font_ids.count; i++) {
 		int32_t glyph = stbtt_FindGlyphIndex(&font_sources[font->font_ids[i]].info, character);
 		if (glyph > 0) {
 			return { glyph, font->font_ids[i] };
@@ -462,20 +462,20 @@ font_glyph_t font_find_glyph(font_t font, char32_t character) {
 int32_t font_add_character(font_t font, char32_t character) {
 	int32_t      index   = -1;
 	font_glyph_t glyph   = font_find_glyph(font, character);
-	if (glyph.idx == 0) return (int32_t)font->character_map.add(character, font->characters['?']);
+	if (glyph.idx == 0) return font->character_map.add(character, font->characters['?']);
 
-	int32_t g_index = (int32_t)font->glyph_map.contains(glyph);
+	int32_t g_index = font->glyph_map.contains(glyph);
 	if (g_index >= 0) {
 		if (character < 128) {
 			font->characters[character] = font->glyph_map.items[g_index];
 			index = -1;
-		} else index = (int32_t)font->character_map.add(character, font->glyph_map.items[g_index]);
+		} else index = font->character_map.add(character, font->glyph_map.items[g_index]);
 	} else {
 		font_char_t ch = font_place_glyph(font, glyph);
 		if (character < 128) {
 			font->characters[character] = ch;
 			index = -1;
-		} else index = (int32_t)font->character_map.add(character, ch);
+		} else index = font->character_map.add(character, ch);
 		font->glyph_map   .add(glyph, ch);
 		font->update_queue.add(glyph);
 	}
@@ -488,7 +488,7 @@ const font_char_t *font_get_glyph(font_t font, char32_t character) {
 	if (character < 128)
 		return &font->characters[character];
 
-	int32_t index = (int32_t)font->character_map.contains(character);
+	int32_t index = font->character_map.contains(character);
 	if (index < 0) {
 		index = font_add_character(font, character);
 	}
@@ -498,7 +498,7 @@ const font_char_t *font_get_glyph(font_t font, char32_t character) {
 ///////////////////////////////////////////
 
 void font_update_cache(font_t font) {
-	for (size_t i = 0; i < font->update_queue.count; i++) {
+	for (int32_t i = 0; i < font->update_queue.count; i++) {
 		font_render_glyph(font, font->update_queue[i], font->glyph_map.get(font->update_queue[i]));
 	}
 	if (font->update_queue.count > 0) {
@@ -510,7 +510,7 @@ void font_update_cache(font_t font) {
 ///////////////////////////////////////////
 
 void font_update_fonts() {
-	for (size_t i = 0; i < font_list.count; i++) {
+	for (int32_t i = 0; i < font_list.count; i++) {
 		font_update_cache(font_list[i]);
 	}
 }
